implemented new BigExp number type to optimize SOFEA some more
BigExp is roughly 30x faster at multiplication than BigDecimal,
but is capped at ~16 decimal digits of precision.
It's really good for doing zSum bound calculations more efficiently,
which is profiling shows is a big bottleneck in SOFEA

also fixed some minor precision issues with SOFEA

+++ test/edu/duke/cs/osprey/sofea/SofeaLab.java
@@ -323,7 +323,7 @@ public class SofeaLab {
 					for (ConfSearch.EnergiedConf econf : confTable.energiedConfs(ConfDB.SortOrder.Energy)) {
 
 						// TEMP: calculate the corrected score
-						BigDecimal zPathUpper = stateInfo.calcZPathUpper(Conf.index(econf.getAssignments()), stateInfo.rcs);
+						BigDecimal zPathUpper = stateInfo.calcZPathUpper(Conf.index(econf.getAssignments()), stateInfo.rcs).toBigDecimal();
 						double correctedScore = sofea.bcalc.freeEnergyPrecise(zPathUpper);
 
 						m.add(sofea.bcalc.calcPrecise(econf.getEnergy()));
